Implemented methods SharedCarUpdater

+++ src/main/java/org/opentripplanner/updater/vehicle_sharing/VehiclePosition.java
@@ -1,6 +1,6 @@
 package org.opentripplanner.updater.vehicle_sharing;
 
 public class VehiclePosition {
-    double longitude;
-    double latitude;
+    float longitude;
+    float latitude;
 }
